Move serialization implementation to separate class
So it can be used as a companion object
